Better ArrayBuffer/Buffer/Uint8Array handling
The type checker complained a bit, and presumably this was fine at some
point, but not anymore. The text encoder and `fs.readFile()` return
`Uint8Arary` objects but potentially over a `SharedArrayBuffer`. I don't
really understand how that works, but the solution is to create another
`Uint8Array` which we know to be over a non-shared `ArrayBuffer`.